v001

+++ src/main/java/com/zyg/exam/service/impl/PaperServiceImpl.java
@@ -25,8 +25,7 @@ public class PaperServiceImpl implements PaperService {
     @Override
     public ResVO selectPaper(PaperDTO paperDTO) {
         List<Object> papers=paperDao.selectByClass(paperDTO).get(0);
-        int total=(int)paperDao.selectByClass(paperDTO).get(1).get(0);
-
+        long total = (long)paperDao.selectByClass(paperDTO).get(1).get(0);
         return new ResVO(papers,total);
     }
 
@@ -43,7 +42,6 @@ public class PaperServiceImpl implements PaperService {
         params.put("pageSize",pageSize);
         List<Object> questions = paperDao.selectQuestion(params).get(0);
         int total=(int)paperDao.selectQuestion(params).get(1).get(0);
-
         return new ResVO(questions,total);
     }
