+++ src/main/java/com/zyg/exam/service/impl/TeachServiceImpl.java
@@ -20,10 +20,8 @@ public class TeachServiceImpl implements TeachService {
     @Override
     public ResVO listTeachInfo(TeachInfoDTO teachInfoDTO) {
         List<Object> teachInfos = teachDao.selectTeachInfo(teachInfoDTO).get(0);
-        int total = (int)teachDao.selectTeachInfo(teachInfoDTO).get(1).get(0);
-
+        long total = (long)teachDao.selectTeachInfo(teachInfoDTO).get(1).get(0);
         return new ResVO(teachInfos,total);
-
     }
 
     @Override
